feat: add ProductFixedItems component
