+++ packages/apps/stock/src/components/organisms/index.js
@@ -35,3 +35,4 @@ export {default as SmallPropertyCard} from './SmallPropertyCard/SmallPropertyCar
 export {default as StockMoveHeader} from './StockMoveHeader/StockMoveHeader';
 export {default as SupplierArrivalCard} from './SupplierArrivalCard/SupplierArrivalCard';
 export {default as SupplierArrivalLineCard} from './SupplierArrivalLineCard/SupplierArrivalLineCard';
+export {default as ProductFixedItems} from './ProductFixedItems/ProductFixedItems';
